moved validation to root dir and password to auth dir
