+++ auth/password.js
